+++ script_checking/GDScriptLocalAnalyzer.gd
@@ -5,5 +5,3 @@ extends RefCounted
 
 @abstract
 func find_any_recursive_function() -> String
-@abstract
-func has_infinite_while_loop() -> bool